Remove module/class name prefix in Swift interface files
See:
https://forums.swift.org/t/frameworkname-is-not-a-member-type-of-frameworkname-errors-inside-swiftinterface/28962
https://bugs.swift.org/browse/SR-898

tl;dr: Swift has problems if the module has the same name as a type
inside that module.
It then treats the name as a type name and prepends the module name
again at type resolution.

We manually (as far as `sed` is "manual") remove that name prefix (`Glean.`)
everywhere and hope it's alright.