Implement the event metric with UniFFI
Because of (current?) UniFFI limitations we need to go a bit of a
special way to keep the existing user-facing API:

In the UDL we define an "internal" event metric.
In Kotlin we build a small shim around it that will have the additional
logic, that knows how to serialize event extra keys into something we
can pass over the FFI.
It can also be made generic then to handle the safe API we currently have.

The downside here is that this again needs to be done in all language
SDKs, but it's limited to those metric types that we can't express in
pure UDL.
UniFFI is gaining support for wrapped types[1], which might be able to
solve this to some extent.
However we still can't express the generic parts in UDL, so some wrapper
will be necessary regardless.

While we're at it we typealias other metrics as well, so they are all
consistenly in the same namespace.
When there's no logic to implement, then that's all we need to do.

[1]: https://mozilla.github.io/uniffi-rs/udl/ext_types_wrapped.html