Add scripts to run Android build in Docker
Our TaskCluster setup runs fully in a Docker container we control.
That makes it easy to replicate the _exact_ setup in a local Docker
setup as well!

So now it becomes as easy as `./taskcluster/docker/linux/runlocally.sh`
to do so.

Note 1: This clones the main branch from Glean. Changing that requires
modifying `./taskcluster/docker/runbuild.sh` for now.

Note 2: this is heavy-weight. The built Docker image clocks in at 15 GB,
loads a whole bunch of stuff from the internet and then the run step
again hits the internet to install Rust over and over again (every time
you run it). This could be optimized but isn't yet.